Add Book Controller

+++ Book_Store.Application/Features/Books/Handlers/Commands/DeleteBookCommandHabdler.cs
@@ -1,16 +1,13 @@
 ﻿using AutoMapper;
-using Book_Store.Application.Features.Books.Requests.Commands;
 using Book_Store.Application.Contracts.Persistence;
+using Book_Store.Application.Features.Books.Requests.Commands;
+using Book_Store.Application.Responses;
+using Book_Store.Domain.Entites;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Book_Store.Application.Features.Books.Handlers.Commands
 {
-    public class DeleteBookCommandHabdler : IRequestHandler<DeleteBookCommand>
+    public class DeleteBookCommandHabdler : IRequestHandler<DeleteBookCommand, BaseCommandResponse>
     {
         private readonly IBookRepository _bookRepository;
         private readonly IMapper _mapper;
@@ -21,11 +18,28 @@ namespace Book_Store.Application.Features.Books.Handlers.Commands
             _mapper = mapper;
         }
 
-        public async Task<Unit> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
+        public async Task<BaseCommandResponse> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
         {
+            var response = new BaseCommandResponse();
+
             var book = await _bookRepository.Get(request.Id);
+
+            if (book is null)
+            {
+                response.Success = false;
+                response.Message = "کتاب یافت نشد.";
+                response.Errors.Add("کتاب یافت نشد.");
+
+                return response;
+            }
+
             await _bookRepository.Delete(book);
-            return Unit.Value;
+
+            response.Success = true;
+            response.Message = "عملیات با موفقیت انجام شد.";
+            response.Id = book.Id;
+
+            return response;
         }
     }
 }
